test(heartbeats): split coverage-smoke into 4 slices + bump --test-timeout to 60s
Coverage-smoke-heartbeats imports 40 heartbeat modules; the per-file
test wrapper was exceeding the 30s --test-timeout in CI (~57s locally).

- Extract HEARTBEATS table + probe() helper into
  _coverage-smoke-heartbeats-shared.mjs (outside the *.test.js glob).
- Split into 4 sibling .test.js files, each importing ~10 modules so
  each file fits under the per-file budget.
- Make the aggregate count test static (just count the HEARTBEATS
  table) instead of re-importing every module.
- Bump npm test/test:watch/test:coverage/test:e2e --test-timeout
  30000 -> 60000 as additional safety for any other heavy
  import-time tests.
