+++ server/tests/coverage-smoke-heartbeats.test.js
@@ -1,116 +1,30 @@
 // server/tests/coverage-smoke-heartbeats.test.js
 //
-// Sprint 34 wave 2 — heartbeat coverage padding.
+// Sprint 34 wave 2 — heartbeat coverage padding (slice 1 of 4).
 //
-// Imports every emergent module with a `run*` export and calls it with a
-// minimal mock context. c8 marks the function as covered once its first
-// line executes; even if the call throws on the first DB query, the
-// function counts AND most of its early scaffolding gets covered.
+// Each *.test.js file runs in its own child process under
+// `node --test`, with its own --test-timeout budget. Imports of the
+// 40-heartbeat-module tree are heavy enough that a single file held
+// the whole list (with module loads on first call) was exceeding
+// the 30 s per-file timeout in CI. Split into 4 slices so each
+// file imports ~10 modules and fits well under the budget.
 //
-// 40 heartbeat modules, 47 total run* exports. Aggregate coverage gain
-// estimated at +0.5pp function coverage; more importantly, this catches
-// any module that breaks at import time or whose run* export signature
-// gets renamed.
-//
-// Reference: docs/security/test-infra-backlog.md item #4 (function
-// coverage thin margin). Anchored to "no regression below current".
+// See _coverage-smoke-heartbeats-shared.mjs for the full
+// HEARTBEATS table + mock-context helper.
 
 import test from "node:test";
 import assert from "node:assert/strict";
+import { HEARTBEATS, probe } from "./_coverage-smoke-heartbeats-shared.mjs";
 
-// Module path → list of run* export names. Verified by Explore-agent sweep
-// against the server/emergent/ tree. Updated when new heartbeat modules
-// land or existing ones get test coverage.
-const HEARTBEATS = [
-  ["../emergent/repair-cortex.js", ["runFullDeploy", "runGuardianCheck", "runProphet", "runSurgeon"]],
-  ["../emergent/sleep-consolidation.js", ["runConsolidation", "runREMPhase"]],
-  ["../emergent/meta-derivation.js", ["runConvergenceCheck", "runMetaDerivationSession"]],
-  ["../emergent/gates.js", ["runAllGates", "runAntiEchoGate"]],
-  ["../emergent/verification-pipeline.js", ["runPipeline"]],
-  ["../emergent/signal-propagation-cycle.js", ["runSignalPropagationCycle"]],
-  ["../emergent/season-cycle.js", ["runSeasonCycle"]],
-  ["../emergent/scenario-engine.js", ["runScenario"]],
-  ["../emergent/research-jobs.js", ["runResearchStep"]],
-  ["../emergent/repair-cycle.js", ["runRepairCycle"]],
-  ["../emergent/procgen-settlement-cycle.js", ["runProcgenSettlementCycle"]],
-  ["../emergent/procedural-npc-spawner.js", ["runProceduralNpcSpawner"]],
-  ["../emergent/population-migration-cycle.js", ["runPopulationMigrationCycle"]],
-  ["../emergent/player-signs-cleanup.js", ["runPlayerSignsCleanup"]],
-  ["../emergent/outcomes.js", ["runWeightLearning"]],
-  ["../emergent/npc-skill-evolve-cycle.js", ["runNpcSkillEvolveCycle"]],
-  ["../emergent/npc-scheme-cycle.js", ["runNpcSchemeCycle"]],
-  ["../emergent/npc-routine-cycle.js", ["runNpcRoutineCycle"]],
-  ["../emergent/npc-marketplace-cycle.js", ["runNpcMarketplaceCycle"]],
-  ["../emergent/npc-economy-cycle.js", ["runNpcEconomyCycle"]],
-  ["../emergent/npc-conversation-initiator.js", ["runNpcConversationInitiator"]],
-  ["../emergent/mount-care-cycle.js", ["runMountCareCycle"]],
-  ["../emergent/lens-learning.js", ["runLensLearningCycle"]],
-  ["../emergent/lattice-quest-cycle.js", ["runLatticeQuestCycle"]],
-  ["../emergent/land-claims-cycle.js", ["runLandClaimsCycle"]],
-  ["../emergent/kingdom-decree-cycle.js", ["runKingdomDecreeCycle"]],
-  ["../emergent/idle-behavior.js", ["runIdleBehavior"]],
-  ["../emergent/hlr-engine.js", ["runHLR"]],
-  ["../emergent/hlm-engine.js", ["runHLMPass"]],
-  ["../emergent/ghost-threads.js", ["runGhostThread"]],
-  ["../emergent/forward-sim-cycle.js", ["runForwardSimCycle"]],
-  ["../emergent/forgetting-engine.js", ["runForgettingCycle"]],
-  ["../emergent/faction-strategy-cycle.js", ["runFactionStrategyCycle"]],
-  ["../emergent/environment-sensor.js", ["runEnvironmentSensor"]],
-  ["../emergent/embodied-dream-cycle.js", ["runEmbodiedDreamCycle"]],
-  ["../emergent/dual-path.js", ["runDualPathSimulation"]],
-  ["../emergent/drift-monitor.js", ["runDriftScan"]],
-  ["../emergent/dream-cycle.js", ["runDreamCycle"]],
-  ["../emergent/deep-health.js", ["runDeepHealthCheck"]],
-  ["../emergent/cross-world-scheme-cycle.js", ["runCrossWorldSchemeCycle"]],
-];
+await probe(test, assert, HEARTBEATS.slice(0, 10));
 
-// Minimal context object — heartbeats expect db, STATE, structuredLog,
-// realtimeEmit. Pass safe no-op shapes; throws inside the handler are
-// caught + ignored (we just need the first line to execute).
-function makeMockCtx() {
-  return {
-    db: null,
-    STATE: {
-      dtus: new Map(),
-      shadowDtus: new Map(),
-      worlds: new Map(),
-      lensArtifacts: new Map(),
-      webhookSecrets: new Map(),
-      _governorTickRunning: false,
-    },
-    structuredLog: () => {},
-    realtimeEmit: () => {},
-    REALTIME: { io: { emit: () => {}, to: () => ({ emit: () => {} }) } },
-    BREAKERS: {},
-    log: () => {},
-  };
-}
-
-for (const [path, runs] of HEARTBEATS) {
-  test(`heartbeat-smoke: ${path} — probe ${runs.join(", ")}`, async () => {
-    const mod = await import(path);
-    for (const name of runs) {
-      const fn = mod[name];
-      assert.equal(typeof fn, "function", `${path} expected export ${name} as function, got ${typeof fn}`);
-      // Call with mock ctx. Throws are expected (db is null) — c8 still
-      // counts the first line as covered.
-      const ctx = makeMockCtx();
-      try { await Promise.resolve(fn(ctx, { db: null, STATE: ctx.STATE })); } catch { /* expected */ }
-      // Some heartbeats take separate args (STATE only, or (db, STATE)).
-      // A second call with different shape covers more entry-line variants.
-      try { await Promise.resolve(fn(ctx.STATE)); } catch { /* expected */ }
-      try { await Promise.resolve(fn()); } catch { /* expected */ }
-    }
-  });
-}
-
-test("heartbeat-smoke: aggregate run* export count ≥ 40", async () => {
+test("heartbeat-smoke: aggregate run* declaration count ≥ 40", () => {
+  // Static count from the HEARTBEATS table — sum of run-export names
+  // across all entries. The per-slice probe() tests above verify that
+  // each declared export actually resolves to a function at import
+  // time, so we don't need to re-import everything here (which would
+  // blow the 30 s file timeout).
   let total = 0;
-  for (const [path, runs] of HEARTBEATS) {
-    const mod = await import(path);
-    for (const name of runs) {
-      if (typeof mod[name] === "function") total++;
-    }
-  }
-  assert.ok(total >= 40, `Only ${total} run* exports found — below 40 floor`);
+  for (const [, runs] of HEARTBEATS) total += runs.length;
+  assert.ok(total >= 40, `Only ${total} run* declarations in HEARTBEATS table — below 40 floor`);
 });
